- Add `named` property to `TestTarget` - Move reference logic from `TesTarget` to an extension
